requirements.txt
boto3==1.10.4
botocore==1.13.4
Django==2.2.2
django-cleanup==4.0.0
django-storages==1.7.2
djangorestframework==3.10.3
docutils==0.15.2
gunicorn==19.9.0
jmespath==0.9.4
numpy==1.17.3
opencv-python==4.1.1.26
Pillow==6.2.0
psycopg2-binary==2.8.4
python-dateutil==2.8.0
pytz==2019.3
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
django-heroku==0.3.1
